+++ _posts/2022-09-25-version-3-1-3.md
@@ -1,0 +1,12 @@
+---
+layout: default
+title: "Version 3.1.3"
+date: 2022-09-25 16:47:00 +0300
+---
+
+## Version 3.1.3
+
+[Link to this release.](https://github.com/vyrjana/DearEIS/releases/tag/3.1.3)
+
+- Fixed bugs that caused the toggling of a plottable series (e.g., a data set or a Kramers-Kronig test result) in the `Plotting` tab to apply the change to the wrong plot under certain circumstances.
+- Fixed bugs that caused a failure to properly adjust the axis limits in cases where the difference between the maximum and minimum values being plotted was zero or all values were zero.
