### 2022-09-28 #4808 @editorialbot
[REVIEW]: DearEIS - A GUI program for analyzing impedance spectra
**Submitting author:** @vyrjana (<a href="http://orcid.org/0000-0001-5779-5201">Ville Yrjänä</a>)
**Repository:** https://github.com/vyrjana/DearEIS
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** 3.4.2
