### 2022-10-25 vyrjana/DearEIS#5 @Aslan-Kosakian
No python executable, only python3
- _DearEIS_ version 3.1.3
- Python version 3.8.10
- OS: Ubuntu 20.04

**Describe the bug**
I am running Ubuntu 20.04, and there is no python executable in the system (only python3). To run the tests (run_tests.sh), I had to change “python” to “python3” in the script. The tests ran fine.

I understand that this can be fixed by introducing a python3 alias for python, but since multiple python versions might be installed in the system and DearEIS uses python3 (using python2 in the test script results in errors), perhaps it would be a good idea to call python3 by default.

**To Reproduce**
Steps to reproduce the behavior:
1. Go to DearEIS/tests/
2. Execute 'run_tests.sh -all'.
