### 2022-10-25 vyrjana/DearEIS#6 @Aslan-Kosakian
build.sh is not executable
- _DearEIS_ version 3.1.3
- Python version 3.8.10
- OS: Ubuntu 20.04

**Describe the bug**
The build.sh script that comes with the software is not executable. If I run “chmod +x build.sh”, then the script complains that it’s been modified and does not run (“Detected uncommitted changes and/or untracked files!”).

**To Reproduce**
Steps to reproduce the behavior:
1. Go to 'DearEIS/'
2. Run 'build.sh'.
