### 2022-10-25 vyrjana/DearEIS#7 @Aslan-Kosakian
generate-api-reference.py missing api_documenter
- _DearEIS_ version 3.1.3
- Python version 3.8.10
- OS: Ubuntu 20.04

**Describe the bug**
If I try to run generate-api-reference.py (in docs), it complains that I don’t have api_documenter.
- This dependency needs to be indicated in the readme.
- This dependency needs to be automatically installed with DearEIS (pip install deareis).
- This dependency cannot be installed with pip. How do I install it?
- Am I supposed to run generate-api-reference.py at all? Its name suggests I am. What does it do?

**To Reproduce**
Steps to reproduce the behavior:
1. Go to 'DearEIS/docs/'
2. Run 'python3 generate-api-reference.py'

**Expected behavior**
Judging by the name, I expect the script to generate an API reference.
