### 2022-10-25 vyrjana/DearEIS#8 @Aslan-Kosakian
LaTeX warning in the API documentation
Desktop:
- DearEIS version 3.1.3
- Python version 3.8.10
- IPython3 version 7.13.0
- OS: Ubuntu 20.04

I get the following warning at the end of the “Generating tables“ section of the "Using the DearEIS API" tutorial (DearEIS/examples/examples.ipynb):

```
/tmp/ipykernel_31743/1140775511.py:6: FutureWarning: In future versions `DataFrame.to_latex` is expected to utilise the base implementation of `Styler.to_latex` for formatting and rendering. The arguments signature may therefore change. It is recommended instead to use `DataFrame.style.to_latex` which also contains additional functionality.
  print(fit.to_dataframe().to_latex(index=False))
/tmp/ipykernel_31743/1140775511.py:13: FutureWarning: In future versions `DataFrame.to_latex` is expected to utilise the base implementation of `Styler.to_latex` for formatting and rendering. The arguments signature may therefore change. It is recommended instead to use `DataFrame.style.to_latex` which also contains additional functionality.
  print(sim.to_dataframe().to_latex(index=False))
```

I'm not sure if this is a problem with my IPython or with DearEIS but wanted to let you know about it anyway, as this warning might turn into an error in the future.

Screenshot:

![DearEIS_API_tutorial_warning](https://user-images.githubusercontent.com/23041345/197858661-a7456518-d536-4872-b7ed-0988820d6c4e.png)
