### 2022-10-25 vyrjana/DearEIS#9 @Aslan-Kosakian
Closing confirmation
Please implement a dialogue window asking the user to confirm closing the application. The opened project is automatically saved and opened the next time the GUI is executed; however, this feature would save the users’ time if they clicked the close button by accident.
