### 2022-10-25 #4808 @Aslan-Kosakian
Hi all. Sorry for the late reply @jgostick and @vyrjana. Here are my comments, questions, and suggestions. Those for which I opened tickets in the DearEIS issue tracker are linked here.

My main concern is that I couldn't find a tutorial on how to actually use the DearEIS GUI. There are good API toturials for DearEIS and pyimpspec, but it would be nice to have a guide for the GUI users too. This would also require a detailed documentation of the GUI's functionality.

### General comments

1. Table 3, line 1 (DearEIS/pyimpspec) - is the last entry supposed to be “GUI, API/CLI”? It currently says “GUI, API/API”. If DearEIS/pyimpspec come with CLI, then please add CLI documentation and tutorials.
2. It is mentioned in the Acknowledgements that DearEIS and pyimpspec (that DearEIS is based on) were developed by the author during their doctoral studies. Since doctoral studies often result in publications and because the JOSS paper requirements include mentions of “any ongoing research projects using the software or recent scholarly publications enabled" by the software, please either a) mention (in the paper) the research and/or publications; or b) confirm (in this discussion, not in the paper) that there is no ongoing research and there are no recent publications based on the software.
3. **([Issue #5](https://github.com/vyrjana/DearEIS/issues/5))** I am running Ubuntu 20.04, and there is no python executable in the system (only python3). To run the tests (run_tests.sh), I had to change “python” to “python3” in the script. The tests ran fine. I understand that this can be fixed by introducing a python3 alias for python, but since multiple python versions might be installed in the system and DearEIS uses python3 (using python2 in the test script results in errors), perhaps it would be a good idea to call python3 by default.
4. **([Issue #6](https://github.com/vyrjana/DearEIS/issues/6))** The build.sh script that comes with the software is not executable. If I run “chmod +x build.sh”, then the script complains that it’s been modified and does not run (“Detected uncommitted changes and/or untracked files!”).
5. **([Issue #7](https://github.com/vyrjana/DearEIS/issues/7))** If I try to run generate-api-reference.py (in docs), it complains that I don’t have api_documenter.
* This dependency needs to be indicated in the readme.
* This dependency needs to be automatically installed with DearEIS (pip install deareis).
* This dependency cannot be installed with pip. How do I install it?
* Am I supposed to run generate-api-reference.py at all? Its name suggests I am. What does it do?

### Statement of need and State of the field

1. Page 1, line 29, about other software solutions: “not all of them are open source”. Please clarify. All packages listed in Table 2 that have their license specified seem to be open-source. What are the licenses of other products? Are they open-source?
2. Page 1, lines 29 and 30, about other software solutions: “in some cases they are not straightforward to use”. Please clarify this statement. Examples could be: no GUI, no application tutorial, non-intuitive interface, etc.
3. Please provide a more detailed overview of the functionality of the existing EIS software and their comparison with DearEIS. For instance:
* Why is it important to include DRT functionality? What can DRT help with in the context of EIS? State that this important functionality is missing from a number of packages (and cite them). For instance, DRT can help identify the number of time constants in the analyzed spectrum and thus decide the number and type of elements to use in the equivalent circuit.
* How do spectrum-fitting capabilities of DearEIS/pyimpspec compare with other software packages? For instance, what are the available circuit elements (does it include CPEs, modified inductance, and Warburg elements, of which there are at least three that I know of?), what are the restrictions (maybe on the maximum number of elements, etc.)?
4. According to Tables 2 and 3, LEVMW/LEVM is the only direct competitor for the proposed software, DearEIS, that has the same functionality (KK, DRT, ECF). While the advantage of DearEIS over other software solutions is clear from Table 3, a statement explaining the advantage of DearEIS over LEVMW/LEVM is needed.
5. Please add [EC-Lab](https://www.biologic.net/support-software/ec-lab-software/), a software from BioLogic, a popular potentiostat manufacturer, to Tables 1-3 and their discussion. EC-Lab is a free (no-cost) and powerful software with EIS and DRT capabilities. It has a GUI and a large number of [technical notes](https://www.biologic.net/document_type/technical-notes/), [application notes](https://www.biologic.net/document_type/application-notes/), and [tutorials](https://www.biologic.net/document_type/tutorials/). What are the DearEIS’ advantages over EC-Lab?
6. Please compare DearEIS/pyimpspec with commercially available software (if there is any). This will help the readers see the perspective and where the open-source EIS software stands at the present time.

### [Online documentation](https://vyrjana.github.io/DearEIS/) referred to by the Readme.md file

1. Item “reading experimental data from several different data formats” is missing a link.

### Readme.md

1. The link to the Scripting section (in Visualization) opens the DearEIS root folder rather than the referred section of the readme file.
2. The readme explains how to install and open DearEIS. However, there is no guide on what to do next. I figured that I could load “example-project.json” in the GUI (this was never mentioned in the guide), but that was it. A step-by-step tutorial showing how to open a sample project, analyze the data, and perform all typical operations is needed. Perhaps it would be best to provide multiple tutorials for different features, e.g., for:
* importing, processing (e.g., averaging), and fitting experimental data (including constructing equivalent circuits using the GUI and via the text line (e.g., [R(RC)(RC)]); this is in the API documentation, but should appear in the tutorial too);
* performing DRT analysis;
* performing KK tests;
* running impedance simulations;
* plotting and figure manipulation;
* saving and exporting data.
Every GUI tab, every menu, every item in the drop-down lists, every parameter/option/feature that appears in the GUI, and all the results (e.g., of running the KK test) should be explained.
3. Where does the example data (example-data.csv) come from? What electrochemical system does it represent? Please add this to the documentation/readme.
4. The included example spectrum is rather simple (two capacitive arcs). It would be worth including multiple examples of spectra typical for different electrochemical systems, e.g., battery (one or more capacitive arcs with a Warburg-like branch at low frequencies), fuel cell (multiple capacitive arcs with an inductive loop at low frequencies), etc.
5. Please provide instructions on the file formatting for the experimental data so that it can be read by DearEIS (see the API comments).
6. (Suggestion) For easier access, it might be a good idea to provide the tutorials as .pdf or .md files that can be readily opened.
7. (Suggestion) Experimentally measured impedance spectra often contain high-frequency inductive component due to external interference. It would be helpful to have a guide on how DearEIS can be used to correct the measured spectra for that interference. For instance, one could fit the high-frequency portion of the spectrum with an equivalent circuit containing an inductive element (or a modified inductance if the inductive tail is not a vertical line - see [this document](https://www.biologic.net/wp-content/uploads/2019/08/battery-eis-modified-inductance-element_electrochemsitry-an42.pdf)) and then subtract the fitted inductive component from the experimental spectrum. It will suffice to show how one could generate an impedance spectrum and subtract it from the measured data using DearEIS GUI.

### "Using the DearEIS API" tutorial (DearEIS/examples/examples.ipynb)

1. Inputs 2 and 3: If matplotlib is imported from deareis as mpl, why do I need to import it again as “import matplotlib”?
2. Input 4: please explain here what the imported classes are and what they do.
3. **([Issue #8](https://github.com/vyrjana/DearEIS/issues/8))** I get the following warning at the end of the “Generating tables“ section:
```
/tmp/ipykernel_31743/1140775511.py:6: FutureWarning: In future versions `DataFrame.to_latex` is expected to utilise the base implementation of `Styler.to_latex` for formatting and rendering. The arguments signature may therefore change. It is recommended instead to use `DataFrame.style.to_latex` which also contains additional functionality.
  print(fit.to_dataframe().to_latex(index=False))
/tmp/ipykernel_31743/1140775511.py:13: FutureWarning: In future versions `DataFrame.to_latex` is expected to utilise the base implementation of `Styler.to_latex` for formatting and rendering. The arguments signature may therefore change. It is recommended instead to use `DataFrame.style.to_latex` which also contains additional functionality.
  print(sim.to_dataframe().to_latex(index=False))
```
4. Say, I want to use my own experimental data. What are the formatting requirements for the experimental data to be imported into DearEIS? This information should include:
- File format
- Type of data that can be read by DearEIS (e.g., real part, imaginary part, frequency - what else?)
- Physical units (Ohm/mOhm/mOhm*cm2/etc. for impedance; Hz or rad/s for frequency)
- Delimiters
- Header names
This information should also be included in the readme file and/or the tutorials (see the Readme comments).
5. The API uses 100% of my CPU (all 8 threads). Is there a way to limit the number of threads DearEIS/pyimpspec is allowed to use?

### Using the [pyimpspec API tutorial](https://github.com/vyrjana/pyimpspec/blob/main/examples/examples.ipynb)

1. The BHT example (input 30) takes a very long time to run. Even if I reduce the number of attempts to 2 and the number of samples to 2, it takes over 5 minutes to run (I stopped it after that) on my 8-thread CPU (and older i7).
2. Input 32 is also taking a long time.

### DearEIS GUI

1. **([Issue #9](https://github.com/vyrjana/DearEIS/issues/9))** Please implement a dialogue window asking the user to confirm closing the application. The opened project is automatically saved and opened the next time the GUI is executed; however, this feature would save the users’ time if they clicked the close button by accident.
