### 2022-10-27 #4808 @vyrjana
Thanks for the comments, @Aslan-Kosakian! I will provide detailed responses to your comments as soon as I can and I will respond to some of the comments in the corresponding issues that you've opened at the DearEIS repository. However, I would first like to clear up a couple of practical things with you, @jgostick, and @mdmurbach.

Is it okay if the manuscript exceeds the 1000 word limit, if that is what it takes to address some of the comments? Some of the points/issues that were raised in @Aslan-Kosakian's comments are a result of my attempt to adhere to the [250-1000 word limit](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain) set forth in the instructions to authors. The manuscript was initially longer and covered some aspects that have been commented on, but I ended up paring down and editing the manuscript to what it is at the moment before submitting.

Is it okay if I make any requested/suggested changes in the `joss-paper` branch during the review process while working as usual on the `main` and `gh-pages` branches and then merge `joss-paper` branch into the `main` and/or `gh-pages` branches at the end of the review process? This would allow me to release updates (e.g., bug fixes) as usual during the review process while keeping the commits made to the `joss-paper` branch related purely to the review process. This would probably make it easier for everyone involved to keep track of what comments have been addressed.

@Aslan-Kosakian and @mdmurbach could then either install the version specified in the first post of this review thread (i.e., version 3.1.2) with the following commands if you need to install it again or on a new system:

`pip install pyimpspec==3.1.2 deareis==3.1.2`

Alternatively, you could clone the `joss-paper` branch of the DearEIS repository and install that branch in development mode using the `--editable` option that `pip` provides.
