### 2022-10-27 #4808 @jgostick
Hi @vyrjana
1. Yes, you can exceed the limit.  It's just a guideline.
2. Yes, you can create a new branch. In fact, you don't even need to merge the paper into the repo...it can remain as an unmerged branch. You do have to merge the code-related changes though.
