### 2022-10-28 vyrjana/DearEIS#9 @vyrjana
This does not seem to be possible at the moment, which is why I implemented automatic saving and recovery of projects with unsaved changes. The GUI framework, Dear PyGui, does have a `set_exit_callback` function but there does not seem to be a way to intercept/cancel the shutdown process like one can with some other frameworks. There is an [open issue](https://github.com/hoffstadt/DearPyGui/issues/227) requesting this feature but that was opened two years ago and as far as I can tell the feature has not yet been implemented.
