### 2022-10-28 6d7e07c
Fixed source of LaTeX-related warnings in the Jupyter notebook
