### 2022-10-28 vyrjana/DearEIS#8 @vyrjana
This is not an issue with DearEIS but rather how the `pandas.DataFrame` object returned by the `FitResult.to_dataframe` method is being used to generate the LaTeX source. I have updated the relevant parts of the Jupyter notebook to get rid of the warnings in [this commit](https://github.com/vyrjana/DearEIS/commit/6d7e07c2ca077b652dc066a82fa615593d5d9388).
