### 2022-10-28 vyrjana/DearEIS#6 @vyrjana
The `build.sh` script can also be executed without making the script executable by passing the file's path to `bash`. The script is also something that very few people would need to run since it is used primarily to build the source and wheel distributions that are uploaded to PyPI.

The script complains about uncommitted changes and/or untracked files specifically as a measure to prevent accidental inclusion of uncommitted changes or untracked files in the source and wheel distributions.

I consider the script to be functioning as intended.
