### 2022-10-28 a347d4a
Updated run_tests.sh to explicitly invoke python3 instead of python
