+++ tests/run_tests.sh
@@ -26,8 +26,8 @@ if [ "$#" -ne 1 ]; then
 	exit
 fi
 if ! [ "$1" == "gui" ]; then
-	python -m unittest discover . -v
+	python3 -m unittest discover . -v
 fi
 if [ "$?" -eq "0" ] && ! [ "$1" == "api" ]; then
-	python -c "from deareis.program import main; from test_gui import setup_tests; setup_tests(); main()"
+	python3 -c "from deareis.program import main; from test_gui import setup_tests; setup_tests(); main()"
 fi
