### 2022-10-28 vyrjana/DearEIS#5 @vyrjana
Yes, changing the script to explicitly invoke `python3` instead of `python` is a reasonable change to make to support distributions that either do not have `python` in the PATH or it points to an earlier version of Python. This change has been implemented in [this commit](https://github.com/vyrjana/DearEIS/commit/a347d4a9b0bc1436a29c7386b7b55cc23c2cbb80).
