### 2022-10-28 vyrjana/DearEIS#7 @vyrjana
That script is used to generate the API documentation that is hosted as part of the project's GitHub pages (i.e., https://vyrjana.github.io/DearEIS/api). The same approach is used for the API documentation of the `pyimpspec` package that DearEIS is based on. Thus, to reduce code duplication, the `api_documenter` package contains the shared code and the script calls functions from that package.

The script is not intended to be executed by a typical end user of the program nor is it required for itself DearEIS to work. As a result, the `api_documenter` package is not available on PyPI and currently it is expected to be obtained separately from the [GitHub repository referenced](https://github.com/vyrjana/python-api-documenter) on line nr. 9 of `generate-api-reference.py`. The package can then be installed locally using by running the `pip install -e .` command from inside the cloned `api_documenter` repository.

I have not previously worked with Git submodules in any of my own projects but that might be a good solution. However, in that case I do need to figure out how to import the `api_documenter` package. It looks like I might need to either modify `sys.path` or use a symlink. I'm not sure if the latter would work as a platform independent solution.
