### 2022-10-28 vyrjana/DearEIS#9 @Aslan-Kosakian
Thank you. Since this is a limitation of Dear PyGui, I'm closing this issue.
