### 2022-10-28 vyrjana/DearEIS#7 @Aslan-Kosakian
If users are not supposed to run generate-api-reference.py, why supply it with DearEIS? Perhaps it could be removed from the user version of the code to avoid confusion.
