### 2022-10-29 469d528
Modified build.sh to explicitly invoke python3 instead of python
