### 2022-10-29 b176be1
Updated documentation for contributing and added optional dependencies for development
