### 2022-10-29 vyrjana/DearEIS#6 @vyrjana
I have updated `build.sh` to use `python3` instead of `python` in [this commit](https://github.com/vyrjana/DearEIS/commit/469d528da4127db56f1612c430df07caf8a02adf). I have also updated `setup.py` to include optional dependencies for development and added a bit of information about this to the "Contributing" section of the readme (see [this commit](https://github.com/vyrjana/DearEIS/commit/b176be1f947d0812fafdb6c3de62e110590c5af0)).
