### 2022-10-29 vyrjana/DearEIS#7 @vyrjana
It is one of the development-related files just like, e.g., `build.sh`. Such files are not included in the distributions uploaded to PyPI, which is where most users would get DearEIS from.
