### 2022-10-30 94aac5f
Updated the documentation and the dependencies relevant for development
