### 2022-10-30 3fbdd6a
Added readme to the docs folder
