### 2022-10-30 b9ab951
Added message to assertions in build.sh
