### 2022-10-30 vyrjana/DearEIS#7 @vyrjana
I have [updated the documentation regarding how to get started with contributing code](https://github.com/vyrjana/DearEIS/commit/94aac5fad542085c509f9b28207cfa3ad1d52df7), [added the `api_documenter` dependency to the project as a Git submodule](https://github.com/vyrjana/DearEIS/commit/39e74cdfcdb44af7e550bf643bc6fc932a9df081), and [added a readme to the docs folder](https://github.com/vyrjana/DearEIS/commit/3fbdd6a5c909bd842004c10855f7bdb7b755c4cd).
