### 2022-10-31 vyrjana/DearEIS#6 @Aslan-Kosakian
Thank you for the changes. The issue can be closed now.
