### 2022-11-01 #4808 @vyrjana
I will add my responses to @Aslan-Kosakian's feedback in chunks. Otherwise it might take me a while to respond to everything at once due to my current schedule.

> My main concern is that I couldn't find a tutorial on how to actually use the DearEIS GUI. There are good API toturials for DearEIS and pyimpspec, but it would be nice to have a guide for the GUI users too. This would also require a detailed documentation of the GUI's functionality.

I've started working on a quickstart guide (or set of tutorials) [here in the `joss-paper-tutorials` branch](https://github.com/vyrjana/DearEIS/tree/joss-paper-tutorials/gh-pages).

> ### General comments
> 1. Table 3, line 1 (DearEIS/pyimpspec) - is the last entry supposed to be “GUI, API/CLI”? It currently says “GUI, API/API”. If DearEIS/pyimpspec come with CLI, then please add CLI documentation and tutorials.

It is correct as it is. DearEIS has a GUI and an API while pyimpspec just has an API. The forward slash is supposed to distinguish DearEIS from pyimpspec, but perhaps that wasn't clear enough and the presentation could be improved by putting them on separate rows.

> 2. It is mentioned in the Acknowledgements that DearEIS and pyimpspec (that DearEIS is based on) were developed by the author during their doctoral studies. Since doctoral studies often result in publications and because the JOSS paper requirements include mentions of “any ongoing research projects using the software or recent scholarly publications enabled" by the software, please either a) mention (in the paper) the research and/or publications; or b) confirm (in this discussion, not in the paper) that there is no ongoing research and there are no recent publications based on the software.

I have not yet had a proper chance to use DearEIS in the research directly related to my own thesis project. I started developing pyimpspec and DearEIS at a relatively late stage in my ongoing doctoral studies at the same time as I was finishing up some mandatory course work that remained. The plan is to use DearEIS in a collaborative project once we get to the stage where we will need to use EIS for material characterization, but I cannot disclose further details about the project at this time due to confidentiality reasons.

DearEIS has been used as a teaching tool these last few weeks in an EIS lab exercise as part of the "Applied Electrochemistry" course at Åbo Akademi University.

> 3. **([Issue #5](https://github.com/vyrjana/DearEIS/issues/5))** I am running Ubuntu 20.04, and there is no python executable in the system (only python3). To run the tests (run_tests.sh), I had to change “python” to “python3” in the script. The tests ran fine. I understand that this can be fixed by introducing a python3 alias for python, but since multiple python versions might be installed in the system and DearEIS uses python3 (using python2 in the test script results in errors), perhaps it would be a good idea to call python3 by default.
> 4. **([Issue #6](https://github.com/vyrjana/DearEIS/issues/6))** The build.sh script that comes with the software is not executable. If I run “chmod +x build.sh”, then the script complains that it’s been modified and does not run (“Detected uncommitted changes and/or untracked files!”).
> 5. **([Issue #7](https://github.com/vyrjana/DearEIS/issues/7))** If I try to run generate-api-reference.py (in docs), it complains that I don’t have api_documenter.
>
> * This dependency needs to be indicated in the readme.
> * This dependency needs to be automatically installed with DearEIS (pip install deareis).
> * This dependency cannot be installed with pip. How do I install it?
> * Am I supposed to run generate-api-reference.py at all? Its name suggests I am. What does it do?
>

These three issues have been addressed (see the respective issues for details).
