+++ _posts/2022-11-01-version-3-2-0.md
@@ -1,0 +1,20 @@
+---
+layout: default
+title: "Version 3.2.0"
+date: 2022-11-01 22:40:00 +0300
+---
+
+## Version 3.2.0
+
+[Link to this release.](https://github.com/vyrjana/DearEIS/releases/tag/3.2.0)
+
+- Added support for calculating the distribution of relaxation times using the `m(RQ)fit` method.
+- Added support (including a keybinding) for loading a simulated spectrum as a data set.
+- Added alternative forms of outputting circuit diagrams as SVG without terminal labels or any labels at all.
+- Updated DearPyGui version requirement to 1.7.3.
+- Updated `DRT analysis` tab to only show settings relevant to the currently chosen method.
+- Updated how the `utility.format_number` function handles the `numpy.nan` value.
+- Updated how the y-axis limits of DRT plots are automatically adjusted.
+- Updated tooltips.
+- Fixed a bug that prevented choosing to use anything but the complex data when performing a DRT analysis using the TR-RBF method.
+- Fixed a bug that was triggered by having too few unmasked data points when performing Kramers-Kronig tests.
