### 2022-11-05 #4808 @jgostick
There is a LOT going on here, and I cannot seem to distill it down to determine if @vyrjana has answered all of @Aslan-Kosakian's questions.  Can either of you give me an update?
