### 2022-11-09 #4808 @mdmurbach
Hi @vyrjana,

Overall a cool UI for doing EIS analysis and seems like a potentially great contribution to the open source impedance analysis space. Aslan-Kosakian's has made some great suggestions for overall improvements, so I don't have too much to add. In fact, my initial comments were similar to theirs in that I struggled a bit to figure out how to use it (despite regularly doing EIS analysis on a pretty regular basis) and I think the project would benefit from a nice introduction example as well as more clear documentation around the different advance feature options (DRT, KK, fitting, etc.).

### Introduction walkthrough (example usage)
In particular, as soon as I opened the app, it wasn't clear to me where to start and if I could directly upload any EIS data directly from a potentiostat or if I needed to reformat it to a specific .csv format (I did eventually realize it's in the pyimpspec data parsing section) and then clicking through the different tabs each took a while to try to parse out what all of the options were for each of the features. In general, once I spent some time playing around with a dataset and trying all of the different features it seems to work pretty well! But I do think it could really benefit from a nice intro/walkthrough to get folks started. I've found these kind of starting-from-the-very-basics examples are super important to growing the number of users beyond a core expert group of users. The gifs that are in the github.io docs are a little hard to follow without any sort of mouse pointer so some written docs with images might also make it easier to get new folks started.

### Functionality documentation
There also isn't any documentation that describes the options for the more advanced features. For example, the DearEIS documentation for DRT points to the pyimpspec docs and those docs point towards a list of different papers. I think it would be great to instead have the docs that are in the [function docstrings](https://github.com/vyrjana/pyimpspec/blob/main/src/pyimpspec/analysis/drt/__init__.py) (nice job documenting those 😄) shown somewhere in the UI documentation.
