+++ joss/paper.bib
@@ -69,6 +69,14 @@
 	url={https://www.iam.kit.edu/et/Lin-KK.php},
 	year={2015}
 }
+@article{schonleber2014method,
+	author={Sch{\"o}nleber, Michael and Klotz, Dino and Ivers-Tiff{\'e}e, E},
+	journal={Electrochimica Acta},
+	pages={20--27},
+	title={A method for improving the robustness of linear Kramers-Kronig validity tests},
+	volume={131},
+	year={2014}
+}
 @misc{knudsen2019pyeis,
 	author={Knudsen, Kristian B},
 	doi={10.5281/zenodo.2535951},
@@ -98,4 +106,96 @@
 	url={http://dx.doi.org/10.1039/D0CP02094J},
 	volume={22},
 	year={2020}
+}
+@misc{zahner,
+	author={{Zahner-Elektrik GmbH \& Co. KG}},
+	journal={GitHub repository},
+	publisher={GitHub},
+	title={{zahner_analysis}},
+	url={https://github.com/Zahner-elektrik/Zahner-Analysis-Python},
+	year={2022}
+}
+@article{matplotlib,
+  author={Hunter, J. D.},
+  title={Matplotlib: A 2D graphics environment},
+  journal={Computing in Science \& Engineering},
+  volume={9},
+  number={3},
+  pages={90--95},
+  abstract={Matplotlib is a 2D graphics package used for Python for
+  application development, interactive scripting, and publication-quality
+  image generation across user interfaces and operating systems.},
+  publisher={IEEE COMPUTER SOC},
+  doi={10.1109/MCSE.2007.55},
+  year={2007}
+}
+@article{lmfit,
+  author={Newville, Matthew and Stensitzki, Till and Allen, Daniel B. and Ingargiola,  Antonino},
+  title={{LMFIT: Non-Linear Least-Square Minimization and Curve-Fitting for Python}},
+  year={2014},
+  publisher={Zenodo},
+  version={0.8.0},
+  doi={10.5281/zenodo.11813},
+  url={https://doi.org/10.5281/zenodo.11813}
+}
+@misc{cdc,
+	author={Boukamp, Bernard A},
+	title={The {CIRCUIT DESCRIPTION CODE} explained},
+	url={https://www.utwente.nl/en/tnw/ims/publications/downloads/},
+	year={2001},
+}
+@article{ciucci2015analysis,
+	author={Ciucci, Francesco and Chen, Chi},
+	doi={10.1016/j.electacta.2015.03.123},
+	journal={Electrochimica Acta},
+	pages={439-454},
+	title={Analysis of Electrochemical Impedance Spectroscopy Data Using the Distribution of Relaxation Times: A Bayesian and Hierarchical Bayesian Approach},
+	volume={167},
+	year={2015}
+}
+@article{effat2017bayesian,
+	author={Effat, Mohammed B and Ciucci, Francesco},
+	doi={10.1016/j.electacta.2017.07.050},
+	journal={Electrochimica Acta},
+	pages={1117-1129},
+	title={Bayesian and Hierarchical Bayesian Based Regularization for Deconvolving the Distribution of Relaxation Times from Electrochemical Impedance Spectroscopy Data},
+	volume={247},
+	year={2017}
+}
+@article{liu2020bayesian,
+	author={Liu, Jiapeng and Wan, Ting Hei and Ciucci, Francesco},
+	doi={10.1016/j.electacta.2020.136864},
+	journal={Electrochimica Acta},
+	pages={136864},
+	title={A Bayesian view on the Hilbert transform and the Kramers-Kronig transform of electrochemical impedance data: Probabilistic estimates and quality scores},
+	volume={357},
+	year={2020}
+}
+@article{boukamp2015fourier,
[…]
