+++ joss/paper.md
@@ -12,6 +12,11 @@ affiliations:
 - name: Åbo Akademi University, Faculty of Science and Engineering, Johan Gadolin Process Chemistry Centre, Laboratory of Molecular Science and Engineering, Henriksgatan 2, FI-20500 Turku (Åbo), Finland
   index: 1
 bibliography: paper.bib
+header-includes:
+	\usepackage{tikz}
+	\usepackage{graphicx}
+	\usepackage{xcolor}
+	\newcommand{\su}[1]{$\rm ^#1$}
 ---
 
 
@@ -19,10 +24,64 @@ bibliography: paper.bib
 
 Electrochemical impedance spectroscopy (EIS) is a technique that is widely used to characterize the properties of materials used in, e.g., batteries and ion-selective electrodes.
 Analyses of the recorded impedance spectra typically involve a data validation step followed by fitting a suitable model to extract quantitative data.
-Analyses of the distribution of relaxation times may also be relevant.
+Analyses of the distribution of relaxation times may also be relevant and can help with choosing an appropriate model.
 DearEIS is a free, open-source, cross-platform program developed for performing such analysis work.
 The primary audience for DearEIS is researchers and engineers.
 However, the program may also prove useful in the context of, e.g., teaching university-level courses on electrochemistry.
+
+
+\begin{figure}
+	\centering
+	\begin{tikzpicture}
+		\draw[draw=none]
+			(0.0cm, 0.0cm) -- 
+			(0.0cm, 9.0cm) -- 
+			(13.5cm, 9.0cm) -- 
+			(13.5cm, 0.0cm) -- 
+			(0.0cm, 0.0cm);
+
+		\def\minnodewidth{0.5cm}
+		\def\minnodeheight{0.5cm}
+		\definecolor{vibrantblue}{HTML}{0077BB}
+		\definecolor{vibrantorange}{HTML}{EE7733}
+		\tikzstyle{process} = [minimum height=\minnodeheight, fill=vibrantblue, opacity=0.2, text opacity=1.0, align=center]
+		\begin{scope}[shift={(1.25cm,6.0cm)}]
+			\node (experimental) at (0.25cm, 1.5cm) [draw, minimum height=\minnodeheight, fill=vibrantorange, opacity=0.2, text opacity=1.0, align=center] {Record \\ spectrum};
+			\node (preprocessing) at (2.75cm, 1.5cm) [draw, process] {Preprocess \\ data};
+			\node (validation) at (5.15cm, 1.5cm) [draw, process] {Validate \\ data};
+			\node (fitting) at (7.75cm, 2.5cm) [draw, process] {Fit equivalent circuit};
+			\node (drt) at (7.75cm, 0.5cm) [draw, process] {Calculate DRT\su{1}};
+			\node (results) at (10.6cm, 1.5cm) [draw, process] {Prepare \\ plots/tables};
+
+
+			\tikzstyle{arrow} = [line width=0.5mm, ->, >=stealth]
+			\tikzstyle{dashedarrow} = [line width=0.5mm, ->, >=stealth, dashed]
+			\draw[arrow] (experimental) -- (preprocessing);
+			\draw[arrow] (preprocessing) -- (validation);
+			\draw[arrow] (validation) |- (fitting);
+			\draw[arrow] (validation) |- (drt);
+			\draw[dashedarrow] (drt) -- (fitting);
+			\draw[arrow] (fitting) -| (results);
+			\draw[arrow] (drt) -| (results);
+
+			\draw[draw=black, line width=0.3mm, dotted]
+				(12.0cm, 0.0cm) --
+				(1.5cm, 0.0cm) --
+				(1.5cm, 3.0cm) --
[…]
