+++ examples/README.md
@@ -4,4 +4,4 @@ The `example-data.csv` file contains the impedance response generated by a circu
 
 The `example-project.json` file contains the example data with and without added noise, some analysis results, and some plots.
 
-The `examples.ipynb` is a Jupyter notebook that demonstrates how one could use the API of DearEIS.
+The `examples.ipynb` is a Jupyter notebook that demonstrates how one could use the API of DearEIS. The `example-data.json` is used in the Jupyter notebook as input for the custom JSON parser.
