### 2022-11-16 f0a57ee
Updated Jupyter notebook - Added comments to imports for various classes in the DearEIS API - Added missing unit for magnitude/modulus in the JSON parser section.
