+++ documentation/API.md
@@ -1,0 +1,5390 @@
+# DearEIS - API reference (3.2.0)
+
+_DearEIS_ is built on top of the `pyimpspec` package.
+See the API reference for `pyimpspec` for information more information about classes and functions that are provided by that package and referenced below (e.g. the `Circuit` class).
+
+**Table of Contents**
+
+- [deareis](#deareis)
+	- [CNLSMethod](#deareiscnlsmethod)
+	- [Capacitor](#deareiscapacitor)
+	- [Circuit](#deareiscircuit)
+		- [get_connections](#deareiscircuitget_connections)
+		- [get_element](#deareiscircuitget_element)
+		- [get_elements](#deareiscircuitget_elements)
+		- [get_label](#deareiscircuitget_label)
+		- [get_parameters](#deareiscircuitget_parameters)
+		- [impedance](#deareiscircuitimpedance)
+		- [impedances](#deareiscircuitimpedances)
+		- [set_label](#deareiscircuitset_label)
+		- [set_parameters](#deareiscircuitset_parameters)
+		- [substitute_element](#deareiscircuitsubstitute_element)
+		- [to_circuitikz](#deareiscircuitto_circuitikz)
+		- [to_drawing](#deareiscircuitto_drawing)
+		- [to_latex](#deareiscircuitto_latex)
+		- [to_stack](#deareiscircuitto_stack)
+		- [to_string](#deareiscircuitto_string)
+		- [to_sympy](#deareiscircuitto_sympy)
+	- [CircuitBuilder](#deareiscircuitbuilder)
+		- [add](#deareiscircuitbuilderadd)
+		- [parallel](#deareiscircuitbuilderparallel)
+		- [series](#deareiscircuitbuilderseries)
+		- [to_circuit](#deareiscircuitbuilderto_circuit)
+		- [to_string](#deareiscircuitbuilderto_string)
+	- [Connection](#deareisconnection)
+		- [contains](#deareisconnectioncontains)
+		- [get_connections](#deareisconnectionget_connections)
+		- [get_element](#deareisconnectionget_element)
+		- [get_elements](#deareisconnectionget_elements)
+		- [get_label](#deareisconnectionget_label)
+		- [get_parameters](#deareisconnectionget_parameters)
+		- [impedance](#deareisconnectionimpedance)
+		- [impedances](#deareisconnectionimpedances)
+		- [set_parameters](#deareisconnectionset_parameters)
+		- [substitute_element](#deareisconnectionsubstitute_element)
+		- [to_latex](#deareisconnectionto_latex)
+		- [to_stack](#deareisconnectionto_stack)
+		- [to_string](#deareisconnectionto_string)
+		- [to_sympy](#deareisconnectionto_sympy)
+	- [ConstantPhaseElement](#deareisconstantphaseelement)
+	- [DRTError](#deareisdrterror)
+	- [DRTMethod](#deareisdrtmethod)
+	- [DRTMode](#deareisdrtmode)
+	- [DRTResult](#deareisdrtresult)
+		- [from_dict](#deareisdrtresultfrom_dict)
+		- [get_bode_data](#deareisdrtresultget_bode_data)
+		- [get_drt_credible_intervals](#deareisdrtresultget_drt_credible_intervals)
+		- [get_drt_data](#deareisdrtresultget_drt_data)
+		- [get_frequency](#deareisdrtresultget_frequency)
+		- [get_gamma](#deareisdrtresultget_gamma)
+		- [get_impedance](#deareisdrtresultget_impedance)
+		- [get_label](#deareisdrtresultget_label)
+		- [get_nyquist_data](#deareisdrtresultget_nyquist_data)
+		- [get_peaks](#deareisdrtresultget_peaks)
[…]
