+++ setup.py
@@ -10,6 +10,20 @@ from os.path import (
 )
 
 
+def update_file(src: str, dst: str):
+    if not exists(src):
+        return
+    src_contents = ""
+    with open(src, "r") as fp:
+        src_contents = fp.read()
+    if exists(dst):
+        with open(dst, "r") as fp:
+            if fp.read() == src_contents:
+                return
+    with open(dst, "w") as fp:
+        fp.write(src_contents)
+
+
 entry_points = {
     "gui_scripts": [
         "deareis = deareis.program:main",
@@ -26,26 +40,16 @@ dependencies = [
     "xdg>=5.1.1",  # Used to figure out where to place config, state, etc. files.
 ]
 
-with open("requirements.txt", "w") as fp:
-    fp.write("\n".join(dependencies))
-
 dev_dependencies = [
     "flake8",
     "setuptools",
 ]
-
-with open("dev-requirements.txt", "w") as fp:
-    fp.write("\n".join(dev_dependencies))
 
 optional_dependencies = {
     "cvxpy": "cvxpy>=1.2.1",  # Used in the DRT calculations (TR-RBF method)
     "kvxopt": "kvxopt>=1.3.0",  # Fork of cvxopt that may provide wheels for additional platforms
     "dev": dev_dependencies,
 }
-
-# The version number defined below is propagated to /src/deareis/version.py
-# when running this script.
-version = "3.2.0"
 
 licenses = []
 for _, _, files in walk("LICENSES"):
@@ -58,49 +62,6 @@ for _, _, files in walk("LICENSES"):
         )
     )
 
-
-def update_file(src: str, dst: str):
-    if not exists(src):
-        return
-    src_contents = ""
-    with open(src, "r") as fp:
-        src_contents = fp.read()
-    if exists(dst):
-        with open(dst, "r") as fp:
-            if fp.read() == src_contents:
-                return
-    with open(dst, "w") as fp:
-        fp.write(src_contents)
-
-
-copyright_notice = ""
-with open("COPYRIGHT") as fp:
-    copyright_notice = fp.read().strip()
-
-with open(join("src", "deareis", "version.py"), "w") as fp:
-    fp.write(f'{copyright_notice}\n\nPACKAGE_VERSION: str = "{version}"')
-
-# The changelog bundled with the package will also be updated when running this script.
-update_file(
-    join("CHANGELOG.md"),
-    join("src", "deareis", "gui", "changelog", "CHANGELOG.md"),
-)
-
-# The licenses bundled with the package will also be updated when running this script.
-update_file(
-    join("LICENSE"),
-    join("src", "deareis", "gui", "licenses", "LICENSE-DearEIS.txt"),
-)
-list(
-    map(
-        lambda _: update_file(
-            _,
-            join("src", "deareis", "gui", "licenses", basename(_)),
-        ),
-        licenses,
-    )
-)
-
 data_files = [
     "COPYRIGHT",
     "CONTRIBUTORS",
@@ -108,40 +69,77 @@ data_files = [
     "src/deareis/gui/changelog/CHANGELOG.md",
 ] + licenses
 
-setup(
-    name="deareis",
-    version=version,
-    author="DearEIS developers",
-    packages=find_packages(where="src"),
-    package_dir={"": "src"},
-    include_package_data=True,
-    data_files=data_files,
-    url="https://vyrjana.github.io/DearEIS",
-    project_urls={
-        "Documentation": "https://vyrjana.github.io/DearEIS/api/",
-        "Source Code": "https://github.com/vyrjana/DearEIS",
-        "Bug Tracker": "https://github.com/vyrjana/DearEIS/issues",
-    },
-    license="GPLv3",
[…]
