+++ joss/paper.md
@@ -23,7 +23,7 @@ header-includes:
 # Summary
 
 Electrochemical impedance spectroscopy (EIS) is a technique that is widely used to characterize the properties of materials used in, e.g., batteries and ion-selective electrodes.
-Analyses of the recorded impedance spectra typically involve a data validation step followed by fitting a suitable model to extract quantitative data.
+Analyses of the recorded impedance spectra typically involve some preprocessing, data validation, and fitting of a suitable model to extract quantitative data.
 Analyses of the distribution of relaxation times may also be relevant and can help with choosing an appropriate model.
 DearEIS is a free, open-source, cross-platform program developed for performing such analysis work.
 The primary audience for DearEIS is researchers and engineers.
@@ -61,19 +61,21 @@ However, the program may also prove useful in the context of, e.g., teaching uni
 			\draw[arrow] (validation) |- (fitting);
 			\draw[arrow] (validation) |- (drt);
 			\draw[dashedarrow] (drt) -- (fitting);
+			\draw[dashedarrow] (fitting.north) |- ++(-.75, .45) -| (preprocessing.north);
 			\draw[arrow] (fitting) -| (results);
 			\draw[arrow] (drt) -| (results);
 
 			\draw[draw=black, line width=0.3mm, dotted]
-				(12.0cm, 0.0cm) --
-				(1.5cm, 0.0cm) --
-				(1.5cm, 3.0cm) --
-				(0.55cm, -0.76cm);
+				(12.00cm,  0.00cm) --
+				( 1.50cm,  0.00cm) --
+				( 1.50cm,  3.55cm) --
+				( 0.57cm, -0.76cm);
+			\draw[draw=none, fill=white] (1.4cm, 3.56cm) rectangle ++(0.2cm, 0.14cm);
 			\draw[draw=black, line width=0.3mm, dotted]
-				(1.5cm, 3.0cm) --
-				(12.0cm, 3.0cm) --
-				(12.0cm, 0.0cm) --
-				(9.44cm, -5.74cm);
+				( 1.5cm,   3.55cm) --
+				(12.00cm,  3.55cm) --
+				(12.00cm,  0.00cm) --
+				( 9.44cm, -5.74cm);
 		\end{scope}
 		\node (screenshot) at (6.25cm, 2.75cm) {\includegraphics[height=5cm, keepaspectratio]{figure-abstract.png}};
 	\end{tikzpicture}
@@ -157,7 +159,7 @@ ZView                  Scribner Associates, Inc.
 
 
 Some of the other software listed in \autoref{tbl:alternatives} are available for free though they are not necessarily also open source (\autoref{tbl:license_comparison}).
-For example, the source code is not available for EIS Spectrum Analyzer, Kramers-Kronig Test, and Lin-KK Tool.
+For example, the source code is not publicly available for EIS Spectrum Analyzer, Kramers-Kronig Test, and Lin-KK Tool.
 The source code for LEVM/LEVMW is included with the binaries, but the source code is not distributed under an open-source license.
 DRTtools requires the user to have licenses for MATLAB and its Optimization Toolbox package, which adds a financial barrier to its use.
 Fortunately, a Python-based port called pyDRTtools is available.
@@ -223,17 +225,17 @@ Most instrument manufacturers bundle their instruments with software that can be
 These two types of software are typically closed source and not publicly available for download (e.g., requiring the purchase of a license or registration of an instrument when making an account).
 Trial versions with some limitations (e.g., an inability to save results) may be publicly available in some cases.
 Restrictive licenses and/or digital rights management technologies (e.g., the USB dongle or the online verification required by RelaxIS) may limit or even prevent distribution of the software to, e.g., colleagues or students.
[…]
