+++ src/deareis/themes.py
@@ -721,12 +721,12 @@ with dpg.theme() as transparent_modal_background:
 # URL theme
 url_theme: int
 with dpg.theme() as url_theme:
-    with dpg.theme_component(dpg.mvInputText, enabled_state=False):
-        dpg.add_theme_color(dpg.mvThemeCol_Text, [255.0, 255.0, 255.0, 255.0])
-        dpg.add_theme_color(dpg.mvThemeCol_Button, [255.0, 255.0, 255.0, 255.0])
-        dpg.add_theme_color(dpg.mvThemeCol_FrameBg, [37.0, 37.0, 38.0, 255.0])
+    with dpg.theme_component(dpg.mvAll, enabled_state=True):
         dpg.add_theme_style(
-            dpg.mvStyleVar_FramePadding, 0.0, 2.0, category=dpg.mvThemeCat_Core
+            dpg.mvStyleVar_ButtonTextAlign,
+            0.0,
+            0.5,
+            category=dpg.mvThemeCat_Core,
         )
 
 
