+++ .github/workflows/test-package.yml
@@ -37,6 +37,7 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         python -m pip install flake8
+        python -m pip install jinja2
         python -m pip install -e .
 
     - name: Lint with flake8
