+++ documentation/API.md
@@ -1,4 +1,4 @@
-# DearEIS - API reference (3.2.0)
+# DearEIS - API reference (3.3.0)
 
 _DearEIS_ is built on top of the `pyimpspec` package.
 See the API reference for `pyimpspec` for information more information about classes and functions that are provided by that package and referenced below (e.g. the `Circuit` class).
