### 2022-11-23 #4808 @vyrjana
Apologies for taking so long to respond to the rest of @Aslan-Kosakian's comments (tagging @mdmurbach to send them a notification). I've made significant changes to the manuscript ([link to most recent PDF](https://github.com/openjournals/joss-papers/blob/joss.04808/joss.04808/10.21105.joss.04808.pdf)) and released new versions of pyimpspec and DearEIS. The tutorials are still a work in progress. I believe I've also managed to figure out the reason for the poor performance that @Aslan-Kosakian reported for the BHT method and fix it (see the end of this post for details).

### **Comment**
> ### Statement of need and State of the field
> 1. Page 1, line 29, about other software solutions: “not all of them are open source”. Please clarify. All packages listed in Table 2 that have their license specified seem to be open-source. What are the licenses of other products? Are they open-source?

### **Response**

EIS Spectrum Analyzer, Kramers-Kronig Test, and Lin-KK Tool are all free to download but the corresponding source code is not publicly available for any of them. The source code for LEVM/LEVMW is included with the program but I've not found any kind of license (OSI-approved or otherwise).

I've added a brief discussion regarding this to the manuscript.

### **Comment**
> 2. Page 1, lines 29 and 30, about other software solutions: “in some cases they are not straightforward to use”. Please clarify this statement. Examples could be: no GUI, no application tutorial, non-intuitive interface, etc.

### **Response**

This is mainly referring to LEVM/LEVMW. Admittedly, I have very limited experience with that software but I do not find the GUI itself intuitive. This is partly due to the heavy use of abbreviations and lack of tooltips. The included documentation is quite comprehensive but one basically has to have it open almost at all times. For example, the circuits are pre-defined and there is some ability to customize them by adjusting the values of the circuit parameters, which have variable meaning depending on the chosen circuit and unfortunately the parameter names in the GUI are not descriptive at all (`p[1]`, `p[2]`, etc.). Fully custom circuits require modifying the source code and there are only two circuits that have been left unimplemented for this very purpose. I'm not sure how easy it would be to add support for having more than two custom circuits at the same time.

I've added some discussion about this to the manuscript. Below are some comments regarding other software but I only included the discussion about LEVM/LEVMW in the manuscript since it is similar to DearEIS in terms of key functionality.

DRT-python-code is a bit clunky to use if you don't have any programming experience. It lacks a proper GUI since it can only be said to have a GUI because it uses matplotlib to show plots. It also does not have a CLI with arguments that could be used to easily process files. Instead one has to modify the script to either include the data or to load the data from a specific path.

Elchemea Analytical's web interface makes extensive use of popup windows when constructing circuits, which most modern web browsers frown upon. This may or may not be an issue depending on the user and whether or not they realize that the GUI is not working properly because the web browser is blocking popup windows.

### **Comment**
> 3. Please provide a more detailed overview of the functionality of the existing EIS software and their comparison with DearEIS. For instance:
>
> * Why is it important to include DRT functionality? What can DRT help with in the context of EIS? State that this important functionality is missing from a number of packages (and cite them). For instance, DRT can help identify the number of time constants in the analyzed spectrum and thus decide the number and type of elements to use in the equivalent circuit.
> * How do spectrum-fitting capabilities of DearEIS/pyimpspec compare with other software packages? For instance, what are the available circuit elements (does it include CPEs, modified inductance, and Warburg elements, of which there are at least three that I know of?), what are the restrictions (maybe on the maximum number of elements, etc.)?
>

### **Response**

I've added a quick mention about the usefulness of DRT in the summary in the manuscript and a short discussion later on in the manuscript. I've also added some examples of functionality beyond KK, DRT, and ECF that some of the other software have. Descriptions of DearEIS' fitting capabilities has also been added to the manuscript. I'm not sure if lmfit, which is the package that is used to implement the fitting, has any limit on the number of parameters, but I have not implemented any kind of limit on the number of circuit elements.

### **Comment**
> 4. According to Tables 2 and 3, LEVMW/LEVM is the only direct competitor for the proposed software, DearEIS, that has the same functionality (KK, DRT, ECF). While the advantage of DearEIS over other software solutions is clear from Table 3, a statement explaining the advantage of DearEIS over LEVMW/LEVM is needed.

### **Response**

I've made changes to the manuscript to clarify, e.g., that:

- LEVM/LEVMW does not have an open-source license (both in the text and in Table 2)
- LEVM/LEVMW may or may not also work natively on platforms other than MS-DOS/Windows
- DearEIS has several additional features besides the ones listed in Table 3

See also my response to your comment regarding lines 29 and 30 on page 1 in the previous version of the manuscript.

As a result of your next two comments, I've added a number of commercial software in the tables. RelaxIS also implements KK, DRT, and ECF. A few comments regarding RelaxIS (e.g., that it is closed-source, commercial software) have also been added to the manuscript.

### **Comment**
> 5. Please add [EC-Lab](https://www.biologic.net/support-software/ec-lab-software/), a software from BioLogic, a popular potentiostat manufacturer, to Tables 1-3 and their discussion. EC-Lab is a free (no-cost) and powerful software with EIS and DRT capabilities. It has a GUI and a large number of [technical notes](https://www.biologic.net/document_type/technical-notes/), [application notes](https://www.biologic.net/document_type/application-notes/), and [tutorials](https://www.biologic.net/document_type/tutorials/). What are the DearEIS’ advantages over EC-Lab?

### **Response**

Obtaining the full version of EC-Lab requires making an account with BioLogic by registering an instrument (a demo version is available). Unfortunately, the BioLogic website is quite sparse on details about the program and its functionality as they relate to analysis of impedance spectra.

See my response to the next comment for information about how DearEIS compares to various commercial software such as EC-Lab.

### **Comment**
> 6. Please compare DearEIS/pyimpspec with commercially available software (if there is any). This will help the readers see the perspective and where the open-source EIS software stands at the present time.

### **Response**

I've added the following commercial software to the tables in the manuscript:

| Name            | Company                             |
| ----------------| ------------------------------------|
| Aftermath       | Pine Research Instrumentation, Inc. |
| EC-Lab          | BioLogic Science Instruments SAS    |
| Echem Analyst   | Gamry Instruments, Inc.             |
| IviumSoft       | Ivium Technologies BV               |
| Nova            | Metrohm AG                          |
| PSTrace         | PalmSens BV                         |
| RelaxIS         | rhd instruments GmbH & Co. KG       |
| Zahner Analysis | Zahner-Elektrik GmbH & Co. KG       |
| ZView           | Scribner Associates, Inc.           |

Some of these are also mentioned specifically in the text when, e.g., comparing functionality.

I have used some of these (Nova, IviumSoft, and Echem Analyst) for analyzing EIS results. As for the others, some of the companies require a purchase of the software license or an instrument for access to the software. Thus, I had to rely on other sources of information (e.g., manuals, product pages, demonstration videos). I've omitted some software that I either could not download or could not find enough information about.

Please let me know if you spot any mistakes (e.g., if you have experience with some of the software that I added to the tables).

### **Comment**
> ### [Online documentation](https://vyrjana.github.io/DearEIS/) referred to by the Readme.md file
> 1. Item “reading experimental data from several different data formats” is missing a link.

### **Response**

I will add a link to the [relevant section of the tutorials (work in progress)](https://github.com/vyrjana/DearEIS/blob/joss-paper-tutorials/gh-pages/tutorials/data.md#supported-file-formats) once the tutorials are ready.

### **Comment**
> ### Readme.md
> 1. The link to the Scripting section (in Visualization) opens the DearEIS root folder rather than the referred section of the readme file.

### **Response**

The link works for me in both Chromium and Firefox.

### **Comment**
> 2. The readme explains how to install and open DearEIS. However, there is no guide on what to do next. I figured that I could load “example-project.json” in the GUI (this was never mentioned in the guide), but that was it. A step-by-step tutorial showing how to open a sample project, analyze the data, and perform all typical operations is needed. Perhaps it would be best to provide multiple tutorials for different features, e.g., for:
>
> * importing, processing (e.g., averaging), and fitting experimental data (including constructing equivalent circuits using the GUI and via the text line (e.g., [R(RC)(RC)]); this is in the API documentation, but should appear in the tutorial too);
> * performing DRT analysis;
> * performing KK tests;
> * running impedance simulations;
> * plotting and figure manipulation;
> * saving and exporting data.
>   Every GUI tab, every menu, every item in the drop-down lists, every parameter/option/feature that appears in the GUI, and all the results (e.g., of running the KK test) should be explained.
>

### **Response**

I am working on writing a set of tutorials dedicated to various features (e.g., creating a project, loading data, masking data points) as I mentioned in [an earlier post](https://github.com/openjournals/joss-reviews/issues/4808#issuecomment-1298187809). There will be a link to [the tutorials](https://github.com/vyrjana/DearEIS/blob/joss-paper-tutorials/gh-pages/tutorials/index.md) in, e.g., the top banner of the GitHub pages (i.e., next to "View On GitHub" and "API documentation"). I tried to provide information about what different settings and buttons do in the program itself using tooltips since external documentation has an unfortunate tendency to become outdated quite easily.

### **Comment**
> 3. Where does the example data (example-data.csv) come from? What electrochemical system does it represent? Please add this to the documentation/readme.

### **Response**

Thanks for pointing this out. It seems that I had forgotten to mention that. The example data is the impedance response of one of the circuits (test circuit 1) that was used in the 1995 article "A linear Kronig-Kramers transform test for immittance data validation" by B.A. Boukamp (DOI:10.1149/1.2044210).

I've added a [readme file](https://github.com/vyrjana/DearEIS/blob/joss-paper/examples/README.md) with this information to the `examples` folder.

### **Comment**
> 4. The included example spectrum is rather simple (two capacitive arcs). It would be worth including multiple examples of spectra typical for different electrochemical systems, e.g., battery (one or more capacitive arcs with a Warburg-like branch at low frequencies), fuel cell (multiple capacitive arcs with an inductive loop at low frequencies), etc.

### **Response**

I'll see what I can do about adding more examples. I've not done any work related to batteries or fuel cells so I do not have spectra for those types of systems.

### **Comment**
> 5. Please provide instructions on the file formatting for the experimental data so that it can be read by DearEIS (see the API comments).

### **Response**

This will be covered in the [relevant tutorial (work in progress)](https://github.com/vyrjana/DearEIS/blob/joss-paper-tutorials/gh-pages/tutorials/data.md#supported-file-formats).

### **Comment**
> 6. (Suggestion) For easier access, it might be a good idea to provide the tutorials as .pdf or .md files that can be readily opened.

### **Response**

This is something I did start implementing before I submitted the initial manuscript. The `docs/generate-api-reference.py` script generates multiple smaller Markdown files that are used for the API documentation that is included in the GitHub pages as well as one large Markdown file. The idea was to use Pandoc to generate a PDF but typesetting ended up being tricky, particularly with functions with many arguments.

I have modified the script to generate the large Markdown file without LaTeX macros and added a [link to it](https://raw.githubusercontent.com/vyrjana/DearEIS/gh-pages/documentation/API.md) on the [index page for the online API documentation](https://vyrjana.github.io/DearEIS/api/).

### **Comment**
> 7. (Suggestion) Experimentally measured impedance spectra often contain high-frequency inductive component due to external interference. It would be helpful to have a guide on how DearEIS can be used to correct the measured spectra for that interference. For instance, one could fit the high-frequency portion of the spectrum with an equivalent circuit containing an inductive element (or a modified inductance if the inductive tail is not a vertical line - see [this document](https://www.biologic.net/wp-content/uploads/2019/08/battery-eis-modified-inductance-element_electrochemsitry-an42.pdf)) and then subtract the fitted inductive component from the experimental spectrum. It will suffice to show how one could generate an impedance spectrum and subtract it from the measured data using DearEIS GUI.
>

### **Response**

This process will be explained in the [relevant tutorial (work in progress)](https://github.com/vyrjana/DearEIS/blob/joss-paper-tutorials/gh-pages/tutorials/data.md#correcting-data-sets). I have made the process a bit easier in the latest version of DearEIS (3.3.0) by making it possible to choose a fit result directly from the `Subtract impedance` window.

### **Comment**
> ### "Using the DearEIS API" tutorial (DearEIS/examples/examples.ipynb)
> 1. Inputs 2 and 3: If matplotlib is imported from deareis as mpl, why do I need to import it again as “import matplotlib”?

### **Response**

The line

`from deareis import mpl`

does not return a reference to matplotlib itself. That line imports a module that contains wrapper functions for plotting data and results using matplotlib as the backend. So `mpl` does not grant direct access to the entire matplotlib package in this context. matplotlib is explicitly imported later on in input 3 in order to configure the default figure size in the Jupyter notebook. The wrapper functions in the `mpl` module do accept matplotlib figures and axes as optional arguments but the wrapper functions create those objects when they are not explicitly provided. Hence, the default figure size is set by first importing matplotlib directly. The alternative would be to create new figures and axes in each input where something is plotted, which defeats the convenience that the wrapper functions are supposed to provide (i.e., the ability to quickly plot stuff with as few lines as possible).

### **Comment**
> 2. Input 4: please explain here what the imported classes are and what they do.

### **Response**

I've added ([see commit](https://github.com/vyrjana/DearEIS/commit/f0a57eec481a3b4f98aca27511faf9aeb26959b0)) brief descriptions of the purposes of those classes and refer the reader to the API documentation for details (e.g., the methods that those classes have).

### **Comment**
> 4. Say, I want to use my own experimental data. What are the formatting requirements for the experimental data to be imported into DearEIS? This information should include:
>
> * File format
> * Type of data that can be read by DearEIS (e.g., real part, imaginary part, frequency - what else?)
> * Physical units (Ohm/mOhm/mOhm*cm2/etc. for impedance; Hz or rad/s for frequency)
> * Delimiters
> * Header names
>   This information should also be included in the readme file and/or the tutorials (see the Readme comments).
>

### **Response**

As mentioned in my response to an earlier comment that touched on this, the formatting requirements will be covered in the [relevant tutorial (work in progress)](https://github.com/vyrjana/DearEIS/blob/joss-paper-tutorials/gh-pages/tutorials/data.md#supported-file-formats).

### **Comment**
> 5. The API uses 100% of my CPU (all 8 threads). Is there a way to limit the number of threads DearEIS/pyimpspec is allowed to use?
>

### **Response**

Yes, but only when using the API rather than the GUI. Some parts of the API make use of multiple processes to improve performance and those functions have a `num_procs` keyword argument that can be used to define the number of processes to use (the value returned by `multiprocessing.cpu_count` is used by default). For example, fitting of equivalent circuits can be done in parallel when multiple combinations of fitting methods and weigths have been chosen.

### **Comment**
> ### Using the [pyimpspec API tutorial](https://github.com/vyrjana/pyimpspec/blob/main/examples/examples.ipynb)
> 1. The BHT example (input 30) takes a very long time to run. Even if I reduce the number of attempts to 2 and the number of samples to 2, it takes over 5 minutes to run (I stopped it after that) on my 8-thread CPU (and older i7).
> 2. Input 32 is also taking a long time.

### **Response**

I have done some investigating and come to the conclusion that the reason is quite likely a combination of:

- the linear algebra library (or libraries) that NumPy is using
- pyimpspec's use of multiple processes

I've been developing and testing primarily on Manjaro where I have installed NumPy via the system package manager.
In this case, NumPy makes use of the CBLAS, BLAS, and LAPACK linear algebra libraries. However, if you install NumPy via PyPI, then NumPy uses the OpenBLAS library, which by default uses multiple threads. The combination of pyimpspec's use of multiprocessing and OpenBLAS' use of multiple threads really hurt the performance of the BHT method.

If you are curious, then you can check which library/libraries NumPy is using by executing the following in your terminal:

`python3 -c "from numpy import show_config; show_config()"`

The output on my desktop system includes:

```
...
blas_info:
    libraries = ['cblas', 'blas', 'blas']
    library_dirs = ['/usr/lib64']
    include_dirs = ['/usr/local/include']
    language = c
    define_macros = [('HAVE_CBLAS', None)]
blas_opt_info:
    define_macros = [('NO_ATLAS_INFO', 1), ('HAVE_CBLAS', None)]
    libraries = ['cblas', 'blas', 'blas']
    library_dirs = ['/usr/lib64']
    include_dirs = ['/usr/local/include']
    language = c
...
lapack_info:
    libraries = ['lapack', 'lapack']
    library_dirs = ['/usr/lib64']
    language = f77
lapack_opt_info:
    libraries = ['lapack', 'lapack', 'cblas', 'blas', 'blas']
    library_dirs = ['/usr/lib64']
    language = c
    define_macros = [('NO_ATLAS_INFO', 1), ('HAVE_CBLAS', None)]
    include_dirs = ['/usr/local/include']
...
```

I did some benchmarking on the following systems (see script at the end of this post):

| CPU              | RAM (GB) | OS                                         | Python | NumPy                      | Linear algebra libraries   |
[…]
