### 2022-11-24 vyrjana/DearEIS#11 @Aslan-Kosakian
Circuit not updated after updating CDC
**Desktop:**
- _DearEIS_ version 3.3.0
- Python version 3.8.10
- OS: Ubuntu 20.0

**Describe the bug**
When a CDC is edited in a window, parsed, and then accepted, the circuit graph in the main DearEIS window (Fitting or Simulation tab) is not updated. It is only updated after “Perform simulation” is clicked.

**To Reproduce**
Steps to reproduce the behavior:
1. Go to the "Fitting" tab and click "Edit" next to the CDC on the left.
2. Modify the CDC, click "Parse". Note the circuit diagram is updated in the "Circuit editor".
3. Click "Accept circuit".
4. Note that the circuit is not updated in the "Fitting" tab.
5. Click "Perform fit". The circuit diagram is updated now.

**Expected behavior**
The graphical representation of the circuit is expected to be automatically updated in step 4 instead of step 5.
