### 2022-11-24 #4808 @Aslan-Kosakian
@vyrjana Thank you for the very comprehensive responses to my comments both in this thread and in the manuscript. The new version of the paper addresses my previous comments fully and in great detail. My only concern in that regard is the discussion of the LEVM/LEVMW shortcomings compared to DearEIS in lines 56-68. The discussion adds details that were lacking in the previous version of the manuscript. However, while the statements made in the new version may be true, the level of detail in the discussion may leave an impression of criticism, especially in the eyes of the LEVM/LEVMW developers and because the rest of the products were discussed only briefly (as they should be). I suggest shrinking this part to a few sentences to highlight only the most critical shortcomings. To avoid negative impressions, perhaps these shortcomings of LEVM/LEVMW could be presented as advantages of DearEIS over that product instead.

### Responses to comments

> Please let me know if you spot any mistakes (e.g., if you have experience with some of the software that I added to the tables).

Unfortunately, I only have some experience with EC-Lab. The information on it in the manuscript seems correct.

> Some parts of the API make use of multiple processes to improve performance and those functions have a num_procs keyword argument that can be used to define the number of processes to use (the value returned by multiprocessing.cpu_count is used by default).

That is fine as long as the user does not want to do anything else with their computer until the API finishes the calculations. Running input 20 of examples.ipynb (perform_test and calculate_drt; this is input 21 in 3.3.0) loads all CPU threads to 100%. I tried to limit the threads to 2 by modifying the code to
```
        max_nfev=1000,
    ),2)
```
and
```
        credible_intervals=False,
    ),2)
```
This did not result in any error but also did not accept my input and still used all 8 threads. This was in 3.1.3. Did I not use num_procs correctly here?

If I use 3.3.0, then the same input (without my modification) does not run and returns the following error:
```
<ipython-input-21-e8ceb7da8201> in <module>
     14     project_ex2.add_test(data, test)
     15
---> 16     drt: DRTResult = deareis.calculate_drt(data, DRTSettings(
     17         method=deareis.DRTMethod.BHT,
     18         rbf_type=deareis.RBFType.GAUSSIAN,

TypeError: __init__() missing 3 required positional arguments: 'circuit', 'W', and 'num_per_decade'
This seems to be an API issue, so I did not create an issue in the DearEIS tracker.
```

Please add instructions on how to change the number of threads to examples.ipynb (and any other parts of documentation where it is relevant). It might be worth explicitly having that argument in all the relevant commands in the document so that the users can change it to a different value that they prefer.

> If you are curious, then you can check which library/libraries NumPy is using by executing the following in your terminal:
> (...)
> The latest versions of pyimpspec (3.2.1) and DearEIS (3.3.0) should fix the performance issues with OpenBLAS without requiring the user to do anything.

My system appears to use the OpenBLAS library. Using pyimpspec 3.2.1, the benchmark you provided runs in a reasonable amount of time (~25-30 seconds on average). Version 3.1.2 runs for much longer (I interrupted it early). So I confirm the fix.

### GUI and tutorial comments

- It might be a good idea to link the tutorials in the Readme document.
- I was not able to load an .mpt file from impedance measurements with a Bio-Logic potentiostat:
```
Traceback (most recent call last):
  File "/home/aslan/.local/lib/python3.8/site-packages/deareis/program/data_sets.py", line 120, in load_data_set_files
    data_sets: List[pyimpspec.DataSet] = pyimpspec.parse_data(path)
  File "/home/aslan/.local/lib/python3.8/site-packages/pyimpspec/data/formats/__init__.py", line 132, in parse_data
    data = func(path, **kwargs)
  File "/home/aslan/.local/lib/python3.8/site-packages/pyimpspec/data/formats/mpt.py", line 51, in parse_mpt
    assert lines.pop(0) == "ec-lab ascii file"
AssertionError
```
I cannot share the file publicly but can send it to you over email for testing purposes (let me know).
- It appears that I can subtract the noisy data set from the ideal data set and the other way around, but I cannot do this with the average data set (subtracting the average or from the average). This is not necessarily a very useful operation with the averaged data set, but if more operations are added in the future, it would be nice to be able to work with the processed data.
- Suggestion for future versions (not for this review): it is sometimes convenient to have a plot of -Im(Z) vs. frequency instead of the Bode plot. It would be great if there was an option in DearEIS to view this graph (it is currently only available in the Plotting tab). Perhaps a toggle could be implemented to show either |Z| or -Im(Z) in other tabs?
- I understand the reason for having a dummy circuit element created between two parallel circuits (e.g., in the tutorial on circuit fitting), but it might be a good idea to explain what it is in the tutorial(s).
- ([Issue #11](https://github.com/vyrjana/DearEIS/issues/11)) When a CDC is edited in a window, parsed, and then accepted, the circuit graph in the main DearEIS window (Fitting or Simulation tab) is not updated. It is only updated after “Perform simulation” is clicked.

Overall, fantastic job!
