### 2022-11-24 #4808 @Aslan-Kosakian
@jgostick I have finished my review and recommend the paper for publication. The only thing that I think needs to be edited before the final publication is the LEVM/LEVMW discussion in the manuscript. The rest of my comments in the previous message are minor, and I do not see any issues preventing new users from successful onboarding with the current version of the software.

@vyrjana Again, great job developing the software and documenting it. DearEIS really fills a niche and has high potential.
