### 2022-11-24 vyrjana/DearEIS#11 @vyrjana
It is working as intended. The `Circuit editor`  window shows what is being constructed while the `Fitting` tab shows the circuit that corresponds to the current result. Perhaps there should be a `Fitted circuit` label above the circuit in the `Fitting`  tab to make this clearer?
