### 2022-11-24 #4808 @vyrjana
@Aslan-Kosakian, thanks again for your comments!

## Comment
> The new version of the paper addresses my previous comments fully and in great detail. My only concern in that regard is the discussion of the LEVM/LEVMW shortcomings compared to DearEIS in lines 56-68. The discussion adds details that were lacking in the previous version of the manuscript. However, while the statements made in the new version may be true, the level of detail in the discussion may leave an impression of criticism, especially in the eyes of the LEVM/LEVMW developers and because the rest of the products were discussed only briefly (as they should be). I suggest shrinking this part to a few sentences to highlight only the most critical shortcomings. To avoid negative impressions, perhaps these shortcomings of LEVM/LEVMW could be presented as advantages of DearEIS over that product instead.

## Response

I was a bit worried about it coming off as too negative. I will try to address this in the next version of the manuscript.

## Comment
> That is fine as long as the user does not want to do anything else with their computer until the API finishes the calculations. Running input 20 of examples.ipynb (perform_test and calculate_drt; this is input 21 in 3.3.0) loads all CPU threads to 100%. I tried to limit the threads to 2 by modifying the code to
>
> ```
>         max_nfev=1000,
>     ),2)
> ```
>
> and
>
> ```
>         credible_intervals=False,
>     ),2)
> ```
>
> This did not result in any error but also did not accept my input and still used all 8 threads. This was in 3.1.3. Did I not use num_procs correctly here?

## Response

That it is technically the correct way to set the number of **processes** that DearEIS (i.e., pyimpspec under the hood) should use though I prefer to explicitly use something like `num_procs=2`. However, OpenBLAS is still trying to use multiple threads of its own in each of those processes unless you tell OpenBLAS otherwise.

The fix that I implemented tries to make it so that most of the time the user should not need to touch the `num_procs` keyword argument nor the `OPENBLAS_NUM_THREADS` environment variable. If OpenBLAS (or MKL) is detected, then pyimpspec will use `num_procs=1` for the BHT method unless told otherwise. Did you try running the input without specifying a value for `num_procs` with DearEIS 3.3.0 and pyimpspec 3.2.1 installed or did you only try it with these latest versions when also specifying `num_procs=2`?

If you do specify a value greater than 1 for `num_procs`, then you should also modify the `OPENBLAS_NUM_THREADS` environemnt variable or else OpenBLAS will use as many threads as it wants per process (i.e., I think you could end up in a situation where you have 2x8=16 threads if you only set `num_procs=2`). It seems that JupyterLab has [a magic for this](https://ipython.readthedocs.io/en/stable/interactive/magics.html#magic-env) (i.e., add `%env OPENBLAS_NUM_THREADS=1` to an input). I'm not sure how well this works compared to setting the environemnt variable in a terminal before running a script or the DearEIS program. I mentioned in a previous post that setting the environemnt variables via Python (i.e., `os.environ["OPENBLAS_NUM_THREADS"] = "1"`) didn't always work consistently during my testing.

## Comment
> If I use 3.3.0, then the same input (without my modification) does not run and returns the following error:
>
> ```
> ---------------------------------------------------------------------------
> TypeError                                 Traceback (most recent call last)
> <ipython-input-21-e8ceb7da8201> in <module>
>      14     project_ex2.add_test(data, test)
>      15
> ---> 16     drt: DRTResult = deareis.calculate_drt(data, DRTSettings(
>      17         method=deareis.DRTMethod.BHT,
>      18         rbf_type=deareis.RBFType.GAUSSIAN,
>
> TypeError: __init__() missing 3 required positional arguments: 'circuit', 'W', and 'num_per_decade'
> This seems to be an API issue, so I did not create an issue in the DearEIS tracker.
> ```

## Response

I either forgot to update this input or failed to merge it into the branch. I'll upload an updated version of the notebook but for now you can also just modify that part of the input to be:

```python
drt: DRTResult = deareis.calculate_drt(data, DRTSettings(
	method=deareis.DRTMethod.BHT,
	rbf_type=deareis.RBFType.GAUSSIAN,
	rbf_shape=deareis.RBFShape.FWHM,
	shape_coeff=0.5,
	derivative_order=1,
	num_samples=2000,
	num_attempts=5,
	maximum_symmetry=0.5,
	# The rest are not necessary/relevant for this method,
	# but these would be always be included when using the GUI
	mode=deareis.DRTMode.COMPLEX,
	lambda_value=-1.0,
	inductance=False,
	credible_intervals=False,
	circuit=None,
	W=0.15,
	num_per_decade=50,
))
```

## Comment
> Please add instructions on how to change the number of threads to examples.ipynb (and any other parts of documentation where it is relevant). It might be worth explicitly having that argument in all the relevant commands in the document so that the users can change it to a different value that they prefer.

## Response

The `num_procs` keyword argument is already present in the function signatures where it is relevant and is documented in the API documentation. I can try to find a good place to add a note about `num_procs` and possible issues when using multithreaded libraries like OpenBLAS.

## Comment
> My system appears to use the OpenBLAS library. Using pyimpspec 3.2.1, the benchmark you provided runs in a reasonable amount of time (~25-30 seconds on average). Version 3.1.2 runs for much longer (I interrupted it early). So I confirm the fix.

## Response

Glad to hear the fix is working.

## Comment
> * It might be a good idea to link the tutorials in the Readme document.

## Response

I will add a link to the tutorials once they are up on the project's GitHub pages.

## Comment
> * I was not able to load an .mpt file from impedance measurements with a Bio-Logic potentiostat:
>
> ```
> Traceback (most recent call last):
>   File "/home/aslan/.local/lib/python3.8/site-packages/deareis/program/data_sets.py", line 120, in load_data_set_files
>     data_sets: List[pyimpspec.DataSet] = pyimpspec.parse_data(path)
>   File "/home/aslan/.local/lib/python3.8/site-packages/pyimpspec/data/formats/__init__.py", line 132, in parse_data
>     data = func(path, **kwargs)
>   File "/home/aslan/.local/lib/python3.8/site-packages/pyimpspec/data/formats/mpt.py", line 51, in parse_mpt
>     assert lines.pop(0) == "ec-lab ascii file"
> AssertionError
> ```
>
> I cannot share the file publicly but can send it to you over email for testing purposes (let me know).

## Response

Perhaps the file structure has changed since at some point and I've implemented the parser on an older version? If you are comfortable doing so, then you can send the file via email (same as my username here + `@abo.fi`). Alternatively, the metadata might be enough (i.e., the lines up until the actual data points start).

## Comment
> * It appears that I can subtract the noisy data set from the ideal data set and the other way around, but I cannot do this with the average data set (subtracting the average or from the average). This is not necessarily a very useful operation with the averaged data set, but if more operations are added in the future, it would be nice to be able to work with the processed data.

## Response

It turns out that I had a bug in the part of the code that filters out data sets with a different number of data points or different frequencies. This bug also affects the filtering in the window for averaging data sets. This will be fixed in the next version.

## Comment
> * Suggestion for future versions (not for this review): it is sometimes convenient to have a plot of -Im(Z) vs. frequency instead of the Bode plot. It would be great if there was an option in DearEIS to view this graph (it is currently only available in the Plotting tab). Perhaps a toggle could be implemented to show either |Z| or -Im(Z) in other tabs?

## Response

I'll make a note of that suggestion. In the meantime, the `DRT analysis` tab does contain a plot with Re(Z) vs f and -Im(Z) vs f.

## Comment
> * I understand the reason for having a dummy circuit element created between two parallel circuits (e.g., in the tutorial on circuit fitting), but it might be a good idea to explain what it is in the tutorial(s).

## Response

I've made a note to add a mention of this to the fitting tutorial. I can also add a tooltip to the `Add dummy` button itself so that one doesn't need to add a dummy node, click on it, and then hover the mouse over the question mark in the `? Dummy node` part on the left-hand side of the *Circuit editor* window to get an idea of what it is for.

## Comment
> * ([Issue #11](https://github.com/vyrjana/DearEIS/issues/11)) When a CDC is edited in a window, parsed, and then accepted, the circuit graph in the main DearEIS window (Fitting or Simulation tab) is not updated. It is only updated after “Perform simulation” is clicked.

## Response

I've responded to this in the opened issue.
