### 2022-11-24 9cbf0d6
Added labels above the circuit previews in the Fitting and Simulation tabs
