### 2022-11-24 vyrjana/DearEIS#11 @vyrjana
I've added labels above the circuit previews in the `Fitting` and `Simulation` tabs in [this commit](https://github.com/vyrjana/DearEIS/commit/9cbf0d6fc61df3a8fe0caea842975dc61abfe369).
