### 2022-11-24 #4808 @jgostick
Hi @vyrjana
Great job addressing all the comments.  I can tell that the release will be much better for it, even though it was quite a lot of effort.  Anyway, before we can move ahead with accepting there are a few more tasks for you to do:

- [ ] Set article DOI with ``@editorialbot set <DOI here> as archive``
- [ ] Do a 'dry run' of acceptance with ``@editorialbot recommend-accept``
