### 2022-11-24 c7a7784
Updated how/when math assets are rendered during startup
