### 2022-11-25 #4808 @Aslan-Kosakian
@vyrjana Thank you for addressing the comments.

> Did you try running the input without specifying a value for num_procs with DearEIS 3.3.0 and pyimpspec 3.2.1 installed or did you only try it with these latest versions when also specifying num_procs=2?

It doesn't run for me at all with DearEIS 3.3.0 and pyimpspec 3.2.1, I included the error message earlier.

> for now you can also just modify that part of the input to be:

That fixed the issue for me, thank you. The code seems to be running much faster now that the issue with OpenBLAS was fixed, so even though DearEIS/pyimpspec uses all threads by default, it doesn't take long. So all good! But thank you for adding more notes on num_procs, it's good to have them just in case.

> you can send the file via email

Sent you the file I tried to load into DearEIS.

I have no follow-up comments for the rest of the topics we discussed.
