+++ tutorials/index.md
@@ -1,0 +1,7 @@
+---
+layout: documentation
+title: Tutorials
+permalink: /tutorials/
+---
+
+TEST
