+++ tutorials/index.md
@@ -4,4 +4,22 @@ title: Tutorials
 permalink: /tutorials/
 ---
 
-TEST
+Here you can find tutorials covering how to use various features of DearEIS and how to use them.
+There are many [hotkeys](settings.md) that can be used for, e.g., more keyboard-centric navigation of the graphical user interface or simply for quickly switching between results.
+
+**IMPORTANT!** The program contains a lot of tooltips with details about what the various settings and buttons do.
+Hover the mouse cursor over labels next to settings/buttons or the buttons themselves to see the tooltips.
+
+
+**Table of Contents**
+
+- [Projects](projects.md)
+- [Processing data](data.md)
+- [Performing Kramers-Kronig tests](kramers-kronig.md)
+- [Performing distribution of relaxation times analyses](drt.md)
+- [Performing equivalent circuit fits](fitting.md)
+- [Performing simulations](simulating.md)
+- [Plotting data](plotting.md)
+- [Program/project settings](settings.md)
+- [Command palette](command-palette.md)
+- [Using the API](https://github.com/vyrjana/DearEIS/blob/main/examples/examples.ipynb)
