### 2022-11-26 e38f6ed
Added button for opening link to tutorials
