+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 3.3.1
 
 - Added labels above the circuit previes in the `Fitting` and `Simulation` tabs to clarify that those correspond to the circuits used in the chosen result rather than what is specified in the settings on the left-hand side.
+- Added button that opens URL to tutorials.
 - Updated how/when certain assets are rendered during startup.
 - Updated how the y-axis limits of DRT plots are automatically adjusted.
 - Updated tooltips.
