+++ _posts/2022-11-26-version-3-4-0.md
@@ -1,0 +1,17 @@
+---
+layout: default
+title: "Version 3.4.0"
+date: 2022-11-26 08:37:00 +0300
+---
+
+## Version 3.4.0
+
+[Link to this release.](https://github.com/vyrjana/DearEIS/releases/tag/3.4.0)
+
+- Added labels above the circuit previes in the `Fitting` and `Simulation` tabs to clarify that those correspond to the circuits used in the chosen result rather than what is specified in the settings on the left-hand side.
+- Added button that opens URL to tutorials.
+- Updated how/when certain assets are rendered during startup.
+- Updated how the y-axis limits of DRT plots are automatically adjusted.
+- Updated tooltips.
+- Fixed a bug that caused data sets to be incorrectly prevented from being selected in the `Average of multiple data sets` and `Subtract impedance` windows.
+- Fixed a bug that caused DRT results to not always load properly.
