### 2022-11-26 c6a1488
Updated minimum version for pyimpspec
