+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 3.4.1
+
+- Updated minimum version for pyimpspec.
+
+
 # 3.4.0
 
 - Added labels above the circuit previes in the `Fitting` and `Simulation` tabs to clarify that those correspond to the circuits used in the chosen result rather than what is specified in the settings on the left-hand side.
