+++ setup.py
@@ -71,7 +71,7 @@ data_files = [
 
 # The version number defined below is propagated to /src/deareis/version.py
 # when running this script.
-version = "3.4.0"
+version = "3.4.1"
 
 if __name__ == "__main__":
     with open("requirements.txt", "w") as fp:
