### 2022-11-26 9cd9eff
Updated API tests - Setting matplotlib backend to get rid of deprecation warnings that are not relevant
