### 2022-11-26 36f62e9
Updated list of contributors - A. Kosakian and M. Murbach helped to improve DearEIS via their peer review of a manuscript about DearEIS (see https://github.com/openjournals/joss-reviews/issues/4808).
