+++ _posts/2022-11-26-version-3-4-1.md
@@ -1,0 +1,11 @@
+---
+layout: default
+title: "Version 3.4.1"
+date: 2022-11-26 11:22:00 +0300
+---
+
+## Version 3.4.1
+
+[Link to this release.](https://github.com/vyrjana/DearEIS/releases/tag/3.4.1)
+
+- Updated minimum version for pyimpspec.
