+++ examples/README.md
@@ -1,0 +1,7 @@
+# Examples
+
+The `example-data.csv` file contains the impedance response generated by a circuit (test circuit 1 or TC-1) that was used in the article "A linear Kronig-Kramers transform test for immittance data validation" by B.A. Boukamp (DOI:10.1149/1.2044210).
+
+The `example-project.json` file contains the example data with and without added noise, some analysis results, and some plots.
+
+The `examples.ipynb` is a Jupyter notebook that demonstrates how one could use the API of DearEIS. The `example-data.json` is used in the Jupyter notebook as input for the custom JSON parser and contains data for the same impedance response as `example-data.csv` except it is given in terms of modulus and phase shift.
