### 2022-11-28 #4808 @vyrjana
@jgostick, I have:

- verified the author and affiliation details
- made a tagged release ([version 3.4.2](https://github.com/vyrjana/DearEIS/releases/tag/3.4.2))
- archived version 3.4.2 on Zenodo  ([10.5281/zenodo.7371986](https://doi.org/10.5281/zenodo.7371986))
- verified that the title and author details in the archived version match those in the manuscript
