+++ _posts/2022-11-28-version-3-4-2.md
@@ -1,0 +1,11 @@
+---
+layout: default
+title: "Version 3.4.2"
+date: 2022-11-28 14:00:00 +0300
+---
+
+## Version 3.4.2
+
+[Link to this release.](https://github.com/vyrjana/DearEIS/releases/tag/3.4.2)
+
+- Updated documentation.
