### 2022-12-05 9e23031
Updated bibliography - Added missing DOI
