+++ joss/paper.bib
@@ -75,7 +75,8 @@
 	pages={20--27},
 	title={A method for improving the robustness of linear Kramers-Kronig validity tests},
 	volume={131},
-	year={2014}
+	year={2014},
+	doi={10.1016/j.electacta.2014.01.034}
 }
 @misc{knudsen2019pyeis,
 	author={Knudsen, Kristian B},
