+++ setup.py
@@ -35,19 +35,20 @@ entry_points = {
 
 dependencies = [
     "dearpygui==1.7.3",  # Used to implement the GUI.
-    "pyimpspec>=3.2.3",  # Used for parsing, fitting, and analyzing impedance spectra.
-    "requests>=2.28.1",  # Used to check package status on PyPI.
-    "xdg>=5.1.1",  # Used to figure out where to place config, state, etc. files.
+    "pyimpspec~=3.2",  # Used for parsing, fitting, and analyzing impedance spectra.
+    "requests~=2.28",  # Used to check package status on PyPI.
+    "xdg~=5.1",  # Used to figure out where to place config, state, etc. files.
 ]
 
 dev_dependencies = [
     "flake8",
     "setuptools",
+    "build",
 ]
 
 optional_dependencies = {
-    "cvxpy": "cvxpy>=1.2.1",  # Used in the DRT calculations (TR-RBF method)
-    "kvxopt": "kvxopt>=1.3.0",  # Fork of cvxopt that may provide wheels for additional platforms
+    "cvxpy": "cvxpy~=1.2",  # Used in the DRT calculations (TR-RBF method)
+    "kvxopt": "kvxopt~=1.3",  # Fork of cvxopt that may provide wheels for additional platforms
     "dev": dev_dependencies,
 }
 
@@ -71,7 +72,7 @@ data_files = [
 
 # The version number defined below is propagated to /src/deareis/version.py
 # when running this script.
-version = "3.4.2"
+version = "3.4.3"
 
 if __name__ == "__main__":
     with open("requirements.txt", "w") as fp:
