+++ tests/test_utility.py
@@ -85,9 +85,9 @@ class TestUtility(TestCase):
         self.assertEqual(format_number(nan), "NaN")
         self.assertEqual(format_number(inf), "INF")
         self.assertEqual(format_number(-inf), "-INF")
-        self.assertEqual(format_number(5.6230364), "5.6    ")
-        self.assertEqual(format_number(5.6230364, decimals=4), "5.6230    ")
-        self.assertEqual(format_number(5.6230364, width=10), "   5.6    ")
+        self.assertEqual(format_number(5.6230364), "5.6")
+        self.assertEqual(format_number(5.6230364, decimals=4), "5.6230")
+        self.assertEqual(format_number(5.6230364, width=10), "       5.6")
         self.assertEqual(format_number(5.6230364, exponent=False), "5.6")
         self.assertEqual(format_number(5.6230364e-5, exponent=False), "0.0")
         self.assertEqual(format_number(5.6230364e-5, exponent=True), "56.2e-06")
