### 2022-12-12 #4808 @Kevin-Mattheus-Moerman
@vyrjana I am the AEiC on this track and here to help process the final steps towards acceptance in JOSS.

I have reviewed the archive and paper. There archive seems in order. I do have some minor points on the paper which require your attention:

- [ ] Currently the paper features `open source` as well as `open-source`, please edit this to be `open source` throughout.
- [ ] You may remove the streetname and number, and the postal codes from the affiliation, i.e. just use Institute, department, town/city, country.
- [ ]  Change `unintuitive` to `not intuitive`.
- [ ] Change `faciliate` to `facilitate`

Please work on the above and let me know when you are finished. Thanks.
