### 2022-12-19 #4808 @vyrjana
@Kevin-Mattheus-Moerman I have been notified about two minor errors in Table 3. I only just saw, after already sending an email to JOSS' admin email address, that the about page states that minor fixes are managed in the GitHub review thread based on requests from the author(s).

The things that need to be corrected are that both DRTtools and pyDRTtools should be marked as capable of performing Kramers-Kronig compliance testing due to the implementation of the Bayesian Hilbert transform method in those software.

How do we proceed?
