### 2022-12-19 #4808 @danielskatz
@vyrjana - the method for doing so is simply for you to make the changes in the paper.md file, and then for @Kevin-Mattheus-Moerman to agree they are minor and then to use the command `@editorialbot reaccept` to regenerate the final pdf and replace it in the archives
