### 2022-12-19 #4808 @vyrjana
Thanks for the information, @danielskatz!

@Kevin-Mattheus-Moerman, I have pushed a [commit with the proposed changes](https://github.com/vyrjana/DearEIS/commit/a3814c955aa031584f230068ffe097ee2d5b6f3e).
